+++ presto-main/src/main/java/com/facebook/presto/operator/scalar/CombineHashFunction.java
@@ -24,7 +24,6 @@ public final class CombineHashFunction
     @SqlType(StandardTypes.BIGINT)
     public static long getHash(@SqlType(StandardTypes.BIGINT) long previousHashValue, @SqlType(StandardTypes.BIGINT) long value)
     {
-        // Hash values are required to be ints
-        return (int) (31 * previousHashValue + value);
+        return (31 * previousHashValue + value);
     }
 }
